Renamed home.js
